+++ src/main/java/onlyoffice/OnlyOfficeEditorServlet.java
@@ -25,7 +25,11 @@ import com.atlassian.confluence.user.AuthenticatedUserThreadLocal;
 import com.atlassian.confluence.user.ConfluenceUser;
 import com.atlassian.confluence.util.velocity.VelocityUtils;
 import com.atlassian.sal.api.message.I18nResolver;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.SerializerProvider;
+import com.fasterxml.jackson.databind.module.SimpleModule;
 import com.onlyoffice.manager.settings.SettingsManager;
 import com.onlyoffice.model.documenteditor.Config;
 import com.onlyoffice.model.documenteditor.config.document.DocumentType;
@@ -34,6 +38,7 @@ import com.onlyoffice.service.documenteditor.config.ConfigService;
 import onlyoffice.managers.auth.AuthContext;
 import com.atlassian.confluence.pages.Attachment;
 import onlyoffice.sdk.manager.document.DocumentManager;
+import onlyoffice.sdk.manager.security.JwtManager;
 import onlyoffice.sdk.manager.url.UrlManager;
 import onlyoffice.utils.attachment.AttachmentUtil;
 import org.apache.log4j.LogManager;
@@ -64,13 +69,14 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
     private final AttachmentUtil attachmentUtil;
     private final ConfigService configService;
     private final SettingsManager settingsManager;
+    private final JwtManager jwtManager;
 
     private final LocaleManager localeManager;
 
     public OnlyOfficeEditorServlet(final I18nResolver i18n, final UrlManager urlManager, final AuthContext authContext,
                                    final DocumentManager documentManager, final AttachmentUtil attachmentUtil,
                                    final ConfigService configService, final SettingsManager settingsManager,
-                                   final LocaleManager localeManager) {
+                                   final JwtManager jwtManager, final LocaleManager localeManager) {
         this.i18n = i18n;
         this.urlManager = urlManager;
         this.authContext = authContext;
@@ -78,6 +84,7 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
         this.attachmentUtil = attachmentUtil;
         this.configService = configService;
         this.settingsManager = settingsManager;
+        this.jwtManager = jwtManager;
         this.localeManager = localeManager;
     }
 
@@ -107,7 +114,7 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
                 String extension = fileExt == null
                         || !fileExt.equals("xlsx")
                         && !fileExt.equals("pptx")
-                        && !fileExt.equals("docxf")
+                        && !fileExt.equals("pdf")
                         ? "docx" : fileExt.trim();
 
                 String name = fileName == null || fileName.equals("")
@@ -178,10 +185,21 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
                         request.getHeader("USER-AGENT")
                 );
 
+                if (modeString != null
+                        && modeString.equals("fillForms")
+                        && config.getDocument().getPermissions().getFillForms()
+                ) {
+                    config.getDocument().getPermissions().setEdit(false);
+                }
+
                 config.getEditorConfig().setLang(localeManager.getLocale(user).toLanguageTag());
                 config.getEditorConfig().setActionLink(actionData);
 
-                ObjectMapper mapper = new ObjectMapper();
+                if (settingsManager.isSecurityEnabled()) {
+                    config.setToken(jwtManager.createToken(config));
+                }
+
+                ObjectMapper mapper = createObjectMapper();
 
                 context.put("request", request);
                 context.put("configAsHtml", mapper.writeValueAsString(config));
@@ -209,5 +227,20 @@ public class OnlyOfficeEditorServlet extends HttpServlet {
         } catch (Exception e) {
             throw new ServletException(e.getMessage(), e);
         }
+    }
+
+    private ObjectMapper createObjectMapper() {
+        ObjectMapper objectMapper = new ObjectMapper();
+        SimpleModule module = new SimpleModule();
+        module.addSerializer(JSONObject.class, new JsonSerializer<JSONObject>() {
+            @Override
+            public void serialize(final JSONObject jsonObject, final JsonGenerator jsonGenerator,
+                                  final SerializerProvider serializerProvider) throws IOException {
+                jsonGenerator.writeObject(jsonObject.toMap());
+            }
+        });
+        objectMapper.registerModule(module);
+
+        return objectMapper;
     }
 }